Refactors `handle_masp_tx`
