+++ wasm/wasm_source/src/tx_transfer.rs
@@ -38,7 +38,7 @@ fn apply_tx(ctx: &mut Ctx, tx_data: Tx) -> TxResult {
         })
         .transpose()?;
     if let Some(shielded) = shielded {
-        token::handle_masp_tx(ctx, &transfer, &shielded)?;
+        token::masp_utils::handle_masp_tx(ctx, &transfer, &shielded)?;
     }
     Ok(())
 }
